requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
pypdf
docarray
requests
beautifulsoup4
